requirements.txt
streamlit
faiss-cpu
groq
PyPDF2
numpy
tqdm
python-dotenv
requests
